src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-root-touch-listener-crash-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
FAIL Unexpected error: Uncaught TypeError: element.createShadowRoot is not a function
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
